# Docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "DeviceBridge",
    "version": "1.0"
  },
  "paths": {
...
    "/devices/{deviceId}/Twin": {
      "get": {
        "tags": [
          "Twin"
        ],
        "summary": "Gets the device twin.",
        "operationId": "GetTwin",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "deviceId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "The device twin.",
            "schema": {
              "type": "object",
              "properties": {
                "twin": {
                  "type": "object",
                  "properties": {
                    "properties": {
                      "type": "object",
                      "properties": {
                        "desired": {
                          "type": "object"
                        },
                        "reported": {
                          "type": "object"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/devices/{deviceId}/Twin/properties/reported": {
      "patch": {
        "tags": [
          "Twin"
        ],
        "summary": "Updates reported properties in the device twin.",
        "description": "Example request:\r\n            \r\n    PATCH /devices/{deviceId}/properties/reported\r\n    {\r\n        \"patch\": {\r\n            \"fanSpeed\": 35,\r\n            \"serial\": \"ABC\"\r\n        }\r\n    }\r\n.",
        "operationId": "UpdateReportedProperties",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "deviceId",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ReportedPropertiesPatch"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Twin updated successfully."
          }
        }
      }
    },
    "/devices/{deviceId}/Twin/properties/desired/sub": {
      "get": {
        "tags": [
          "Twin"
        ],
        "summary": "Gets the current desired property change subscription for a device.",
        "operationId": "GetDesiredPropertiesSubscription",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "deviceId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "The current desired property change subscription.",
            "schema": {
              "$ref": "#/definitions/DeviceSubscriptionWithStatus"
            }
          },
          "404": {
            "description": "If a subscription doesn't exist."
          }
        }
      },
      "put": {
        "tags": [
          "Twin"
        ],
        "summary": "Creates or updates the current desired property change subscription for a device.",
        "description": "When the device receives a new desired property change from IoTHub, the service will send an event to the desired callback URL.\r\n            \r\n    Example event:\r\n    {\r\n        \"eventType\": \"string\",\r\n        \"deviceId\": \"string\",\r\n        \"deviceReceivedAt\": \"2020-12-04T01:06:14.251Z\",\r\n        \"desiredProperties\": {\r\n            \"prop1\": \"string\",\r\n            \"prop2\": 12,\r\n            \"prop3\": {},\r\n        }\r\n    }\r\n.",
        "operationId": "CreateOrUpdateDesiredPropertiesSubscription",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "deviceId",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/SubscriptionCreateOrUpdateBody"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The created or updated C2D message subscription.",
            "schema": {
              "$ref": "#/definitions/DeviceSubscriptionWithStatus"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Twin"
        ],
        "summary": "Deletes the current desired property change subscription for a device.",
        "operationId": "DeleteDesiredPropertiesSubscription",
        "parameters": [
          {
            "in": "path",
            "name": "deviceId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "Subscription deleted successfully."
          }
        }
      }
    }
  },
  "definitions": {
...
    "SubscriptionCreateOrUpdateBody": {
      "required": [
        "callbackUrl"
      ],
      "type": "object",
      "properties": {
        "callbackUrl": {
          "type": "string"
        }
      }
    },
    "DeviceSubscriptionWithStatus": {
      "type": "object",
      "properties": {
        "deviceId": {
          "type": "string"
        },
        "subscriptionType": {
          "type": "string"
        },
        "callbackUrl": {
          "type": "string"
        },
        "createdAt": {
          "format": "date-time",
          "type": "string"
        },
        "status": {
          "type": "string"
        }
      }
    },
...
    "ReportedPropertiesPatch": {
      "required": [
        "patch"
      ],
      "type": "object",
      "properties": {
        "patch": {
          "type": "object",
          "additionalProperties": { }
        }
      }
    }
  }
}